back-end/requirements.txt
absl-py==1.4.0
APScheduler==3.10.3
astunparse==1.6.3
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
filelock==3.12.2
Flask==2.3.2
flatbuffers==23.5.26
fsspec==2023.6.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
JPype1==1.4.1
keras==2.13.1
konlpy==0.6.0
libclang==16.0.6
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
soupsieve==2.4.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-addons==0.21.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
threadpoolctl==3.2.0
tokenizers==0.13.3
tqdm==4.66.1
transformers==4.31.0
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
Werkzeug==2.3.7
wrapt==1.15.0
zipp==3.16.2
